move Badges to own namespace in preparation for Emote API support
