+++ TwitchLib.Api.Helix.Models/Chat/Badges/EmoteSet.cs
@@ -1,9 +1,6 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace TwitchLib.Api.Helix.Models.Chat
+namespace TwitchLib.Api.Helix.Models.Chat.Badges
 {
     public class EmoteSet
     {
