+++ TwitchLib.Api.Helix.Models/Chat/Badges/GetChannelChatBadges/GetChannelChatBadgesResponse.cs
@@ -1,9 +1,6 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace TwitchLib.Api.Helix.Models.Chat.GetChannelChatBadges
+namespace TwitchLib.Api.Helix.Models.Chat.Badges.GetChannelChatBadges
 {
     public class GetChannelChatBadgesResponse
     {
