+++ TwitchLib.Api.Helix.Models/Chat/Badges/GetGlobalChatBadges/GetGlobalChatBadgesResponse.cs
@@ -1,9 +1,6 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace TwitchLib.Api.Helix.Models.Chat.GetGlobalChatBadges
+namespace TwitchLib.Api.Helix.Models.Chat.Badges.GetGlobalChatBadges
 {
     public class GetGlobalChatBadgesResponse
     {
