+++ TwitchLib.Api.Helix.Models/Chat/Badges/Version.cs
@@ -1,9 +1,6 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace TwitchLib.Api.Helix.Models.Chat
+namespace TwitchLib.Api.Helix.Models.Chat.Badges
 {
     public class Version
     {
